+++ dbt/include/firebolt/macros/adapters.sql
@@ -46,11 +46,6 @@
 {%- endmacro %}
 
 
---TODO
----add column to table
---  ALTER TABLE {{ relation }}
---  ADD COLUMN modify column {{ adapter.quote(column_name) }}
----    {{ new_column_type }} {{ on_cluster_clause(label="on cluster") }}
 {% macro firebolt__alter_column_type(relation, column_name, new_column_type) -%}
     {# stub #}
     {% call statement('alter_column_type') %}
@@ -73,14 +68,19 @@
                                create_statement,
                                spine_col,
                                other_col) -%}
-  {{ create_statement }} "{{ index_name }}" ON {{ relation }} (
-      {{ spine_col }},
-      {% if other_col is iterable and other_col is not string -%}
-          {{ other_col | join(', ') }}
-      {%- else -%}
-          {{ other_col }}
-      {%- endif -%}
-      );
+    {# Write SQL for generating a join or aggregating index. #}
+    {{ create_statement }} "{{ index_name }}" ON {{ relation }} (
+        {% if spine_col is iterable and spine_col is not string -%}
+            {{ spine_col | join(', ') }},
+        {% else -%}
+            {{ spine_col }},
+        {% endif -%}
+        {% if other_col is iterable and other_col is not string -%}
+            {{ other_col | join(', ') }}
+        {%- else -%}
+            {{ other_col }}
+        {%- endif -%}
+    );
 {%- endmacro %}
 
 
@@ -92,18 +92,22 @@
 
 
 {% macro firebolt__get_create_index_sql(relation, index_dict) -%}
+    {# Parse index inputs and send parsed input to make_create_index_sql #}
     {%- set index_config = adapter.parse_index(index_dict) -%}
-    {%- set index_name = index_config.render(relation) -%}
-    {%- set index_type = index_config.type | upper -%}
-
+    {%- set index_name = index_config.render_name(relation) -%}
+    {%- set index_type = index_config.index_type | upper -%}
     {%- if index_type == "JOIN" -%}
-        {{ make_create_index_sql(
-          relation, index_name, "CREATE JOIN INDEX",
-          index_config.join_column, index_config.dimension_column) }}
+        {{ make_create_index_sql(relation,
+                                 index_name,
+                                 "CREATE JOIN INDEX",
+                                 index_config.join_column,
+                                 index_config.dimension_column) }}
     {%- elif index_type == "AGGREGATING" -%}
-        {{ make_create_index_sql(
-          relation, index_name, "CREATE AND GENERATE AGGREGATING INDEX",
-          index_config.key_column, index_config.aggregation) }}
+        {{ make_create_index_sql(relation,
+                                 index_name,
+                                 "CREATE AND GENERATE AGGREGATING INDEX",
+                                 index_config.key_column,
+                                 index_config.aggregation) }}
     {%- endif -%}
 {%- endmacro %}
 
@@ -124,8 +128,6 @@
 
 
 {% macro firebolt__list_relations_without_caching(schema_relation) %}
-    {# TODO: schema_relation is ??
-             What does "without caching" mean in this context? #}
     {% call statement('list_tables_without_caching', fetch_result=True) -%}
         SELECT '{{ schema_relation.database }}' AS "database",
                table_name AS "name",