+++ CHANGELOG.md
@@ -4,8 +4,13 @@
 
 ### Under the hood
 
+- dbt-firebolt now supports aggregating indexes with multiple-column keys. 
 - Bug fix to check for `data_type` and `regex` fields when necessary on external tables.
 - Changed default behavior on external table insert to `DROP IF EXISTS`.
+
+### Breaking changes
+
+- Models with aggregating and join indexes now require `index_type` field in index config blocks rather than `type`.
 
 ## v.1.0.1